add support for upper bounding the SVD composition
when truncating an SVD and then composing it into A_trunc, some values in
A_trunc may be larger than A. This is fine, except when you're representing
image data. values greater than the acceptable bit depth will be cleared (or
something weird; couldn't quite figure it out). this has the effect of creating
an image with strange blotches for pixels that had this problem.

it's possible that this behavior is unique in how RMagick created colors from
channels.

this commit solves this by giving the Svd.compose function an options argument.
if you set :upper_bound to a Numeric, all values greater than that will be set
to that number in the matrix